Saturdaty 26/12/15, 10:12am.
This is the first commit of the assignment, function working properly.

The function works fine. In this commit I included the "trycatch" command
to see if I was able to handle errors due to a matrix that R is not able
to solve (not a problem with my functions, but with "solve").